Added 1,2 and 4 byte literals. Pattern matching for simple literal patterns

+++ frontends/my_tests/ExploratoryTest.scala
@@ -3,7 +3,7 @@
 object ExploratoryTest {
 
 
-  def foo(baz: BigInt) = baz match {
+  def foo(baz: Int) = baz match {
     case a @ 1 => "Romain"
     case b @ 2 => "Stevan"
   }
